refactor(web): split the drawer monolith into a folder behind its barrel (Tier-2 U13)
